TSBLoader_adv (command line tool) updated to version 1.0.5 New GUI Tool for using TSB Loader v 1.0.1
Version 1.0.5 include several extensions specific to Seed Robotics use of the products, such as the DynamixelCommandGenerator. This is used for the new 'live' bootloader activation mode. In this scenario, a device actively running a firmware, will jump to the bootloader (soft activation) by sending a specific Dynamixel command. For general uses, we recommend users stay with the 'cold' activation mode (the one used by default, where you power cycle the unit)